Revert "forgot bulk delete"
This reverts commit 1b4fb999ee07b358ee6e1af9efb8981b84f83af1.
