+++ src/stores/MessageStore.js
@@ -1,8 +1,7 @@
 const DataStore = require('./DataStore');
 const Collection = require('../util/Collection');
 const Message = require('../structures/Message');
-const { Error, TypeError } = require('../errors');
-const Snowflake = require('../util/Snowflake');
+const { Error } = require('../errors');
 
 /**
  * Stores messages for text-based channels.
@@ -66,51 +65,6 @@ class MessageStore extends DataStore {
       for (const message of data) messages.set(message.id, this.add(message));
       return messages;
     });
-  }
-
-  /**
-   * Bulk deletes given messages that are newer than two weeks.
-   * <warn>This is only available when using a bot account.</warn>
-   * @param {Collection<Snowflake, Message>|Message[]|Snowflake[]|number} messages
-   * Messages or number of messages to delete
-   * @param {boolean} [filterOld=false] Filter messages to remove those which are older than two weeks automatically
-   * @returns {Promise<Collection<Snowflake, Message>>} Deleted messages
-   * @example
-   * // Bulk delete messages
-   * channel.messages.bulkDelete(5)
-   *   .then(messages => console.log(`Deleted ${messages.size} messages`))
-   *   .catch(console.error);
-   */
-  async bulkDelete(messages, filterOld = false) {
-    if (messages instanceof Array || messages instanceof Collection) {
-      let messageIDs = messages instanceof Collection ? messages.keyArray() : messages.map(m => m.id || m);
-      if (filterOld) {
-        messageIDs = messageIDs.filter(id =>
-          Date.now() - Snowflake.deconstruct(id).date.getTime() < 1209600000
-        );
-      }
-      if (messageIDs.length === 0) return new Collection();
-      if (messageIDs.length === 1) {
-        await this.client.api.channels(this.channel.id).messages(messageIDs[0]).delete();
-        const message = this.client.actions.MessageDelete.handle({
-          channel_id: this.channel.id,
-          id: messageIDs[0],
-        }).message;
-        if (message) return new Collection([[message.id, message]]);
-        return new Collection();
-      }
-      await this.client.api.channels[this.channel.id].messages['bulk-delete']
-        .post({ data: { messages: messageIDs } });
-      return this.client.actions.MessageDeleteBulk.handle({
-        channel_id: this.channel.id,
-        ids: messageIDs,
-      }).messages;
-    }
-    if (!isNaN(messages)) {
-      const msgs = await this.fetch({ limit: messages });
-      return this.bulkDelete(msgs, filterOld);
-    }
-    throw new TypeError('MESSAGE_BULK_DELETE_TYPE');
   }
 
   _fetchId(messageID) {
